requirements.txt
tensorflow==2.13.0
flask==2.2.3
scikit-learn==1.2.2
pillow==8.4.0
matplotlib==3.7.1
uvicorn==0.22.0
fastapi==0.95.1
seaborn==0.11.2